Fixed 'undefined' in Carousel cards
'Undefined' was showing up as a message for Genesis and Team
members who didn't have a defined quote in Contentful.
Made the quote field an option(string) type in ContentTypes.TeamMember
and ContentTypes.GenesisMember. Added a switch statement that checks
for for a quote field and conditionally displays a message now.